3094 don't log activity entries at updatelog
